--- src/main/java/gov/energy/nbc/car/bo/mongodb/multipleCellSchemaApproach/m_DatasetBO.java
@@ -1,202 +1,0 @@
-package gov.energy.nbc.car.bo.mongodb.multipleCellSchemaApproach;
-
-import com.mongodb.util.JSON;
-import gov.energy.nbc.car.bo.FileStorageBO;
-import gov.energy.nbc.car.bo.exception.DeletionFailure;
-import gov.energy.nbc.car.bo.mongodb.AbsDatasetBO;
-import gov.energy.nbc.car.dao.IDatasetDAO;
-import gov.energy.nbc.car.utilities.FileAsRawBytes;
-import gov.energy.nbc.car.dao.dto.IDeleteResults;
-import gov.energy.nbc.car.dao.mongodb.DAOUtilities;
-import gov.energy.nbc.car.dao.mongodb.multipleCellCollectionsApproach.m_DatasetDAO;
-import gov.energy.nbc.car.model.IDatasetDocument;
-import gov.energy.nbc.car.model.IMetadata;
-import gov.energy.nbc.car.model.IRowCollection;
-import gov.energy.nbc.car.model.IStoredFile;
-import gov.energy.nbc.car.model.mongodb.common.Metadata;
-import gov.energy.nbc.car.model.mongodb.common.StoredFile;
-import gov.energy.nbc.car.model.mongodb.document.DatasetDocument;
-import gov.energy.nbc.car.settings.ISettings;
-import gov.energy.nbc.car.utilities.PerformanceLogger;
-import gov.energy.nbc.car.utilities.fileReader.DatasetReader_AllFileTypes;
-import gov.energy.nbc.car.utilities.fileReader.IDatasetReader_AllFileTypes;
-import gov.energy.nbc.car.utilities.fileReader.dto.RowCollection;
-import gov.energy.nbc.car.utilities.fileReader.exception.InvalidValueFoundInHeader;
-import gov.energy.nbc.car.utilities.fileReader.exception.UnsupportedFileExtension;
-import org.apache.log4j.Logger;
-import org.bson.Document;
-import org.bson.types.ObjectId;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-
-public class m_DatasetBO extends AbsDatasetBO {
-
-    Logger log = Logger.getLogger(this.getClass());
-
-    protected m_DatasetDAO datasetDAO;
-    protected FileStorageBO fileStorageBO;
-
-    protected IDatasetReader_AllFileTypes generalFileReader;
-
-    public m_DatasetBO(ISettings settings) {
-
-        super(settings);
-
-        datasetDAO = new m_DatasetDAO(settings);
-        fileStorageBO = new FileStorageBO(settings);
-
-        generalFileReader = new DatasetReader_AllFileTypes();
-    }
-
-    @Override
-    public ObjectId addDataset(
-            String dataCategory,
-            Date submissionDate,
-            String submitter,
-            String projectName,
-            String chargeNumber,
-            String comments,
-            gov.energy.nbc.car.dao.dto.StoredFile sourceDocument,
-            String nameOfSubdocumentContainingDataIfApplicable,
-            List<gov.energy.nbc.car.dao.dto.StoredFile> attachmentFiles)
-            throws UnsupportedFileExtension, InvalidValueFoundInHeader {
-
-        File storedFile = fileStorageBO.getFile(sourceDocument.storageLocation);
-        RowCollection dataUpload = generalFileReader.extractDataFromFile(storedFile, nameOfSubdocumentContainingDataIfApplicable, -1);
-        IRowCollection rowCollection = new gov.energy.nbc.car.model.mongodb.common.RowCollection(dataUpload.columnNames, dataUpload.rowData);
-
-        List<IStoredFile> attachments = new ArrayList();
-        if (attachmentFiles != null) {
-            for (gov.energy.nbc.car.dao.dto.StoredFile attachmentFile : attachmentFiles) {
-                attachments.add(new StoredFile(attachmentFile.originalFileName, attachmentFile.storageLocation));
-            }
-        }
-
-        IDatasetDocument datasetDocument = new DatasetDocument(
-                dataCategory,
-                submissionDate,
-                submitter,
-                chargeNumber,
-                projectName,
-                comments,
-                new StoredFile(sourceDocument.originalFileName, sourceDocument.storageLocation),
-                nameOfSubdocumentContainingDataIfApplicable,
-                attachments);
-
-        ObjectId objectId = getDatasetDAO().add(datasetDocument, rowCollection);
-
-        return objectId;
-    }
-
-    @Override
-    public String getDataset(String datasetId) {
-
-        IDatasetDocument datasetDocument = getDatasetDAO().getDataset(datasetId);
-        if (datasetDocument == null) { return null; }
-
-        String jsonOut = DAOUtilities.serialize(datasetDocument);
-        return jsonOut;
-    }
-
-    @Override
-    public String getAllDatasets() {
-
-        Iterable<Document> datasets = getDatasetDAO().getAll();
-
-        String jsonOut = DAOUtilities.serialize(datasets);
-        return jsonOut;
-    }
-
-    @Override
-    public long removeDataset(String datasetId) throws DeletionFailure {
-
-        m_DatasetDAO datasetDAO = getDatasetDAO();
-        IDatasetDocument datasetDocument = datasetDAO.getDataset(datasetId);
-
-        String storageLocation = datasetDocument.getMetadata().getSourceDocument().getStorageLocation();
-        try {
-            fileStorageBO.moveFilesToRemovedFilesLocation(storageLocation);
-        } catch (IOException e) {
-            log.warn(e);
-        }
-
-        IDeleteResults deleteResults = datasetDAO.delete(datasetId);
-
-        return deleteResults.getDeletedCount();
-    }
-
-    @Override
-    public String addDataset(
-            String dataCategory,
-            Date submissionDate,
-            String submitter,
-            String projectName,
-            String chargeNumber,
-            String comments,
-            FileAsRawBytes sourceDocument,
-            String nameOfSubdocumentContainingDataIfApplicable,
-            List<FileAsRawBytes> attachmentFiles)
-            throws UnsupportedFileExtension, InvalidValueFoundInHeader {
-
-        Date timestamp = new Date();
-
-        gov.energy.nbc.car.dao.dto.StoredFile theDataFileThatWasStored = fileStorageBO.saveFile(timestamp, "", sourceDocument);
-
-        List<gov.energy.nbc.car.dao.dto.StoredFile> theAttachmentsThatWereStored = new ArrayList();
-
-        for (FileAsRawBytes attachmentFile : attachmentFiles) {
-            theAttachmentsThatWereStored.add(fileStorageBO.saveFile(timestamp, "attachments", attachmentFile));
-        }
-
-        ObjectId objectId = addDataset(
-                dataCategory,
-                submissionDate,
-                submitter,
-                projectName,
-                chargeNumber,
-                comments,
-                theDataFileThatWasStored,
-                nameOfSubdocumentContainingDataIfApplicable,
-                theAttachmentsThatWereStored);
-
-        return JSON.serialize(objectId);
-    }
-
-    @Override
-    public String addDataset(String metadataJson,
-                             File file,
-                             String nameOfSubdocumentContainingDataIfApplicable)
-            throws UnsupportedFileExtension, InvalidValueFoundInHeader {
-
-        IDatasetDAO datasetDAO = getDatasetDAO();
-
-        try {
-            RowCollection dataUpload = generalFileReader.extractDataFromDataset(file, nameOfSubdocumentContainingDataIfApplicable);
-            IRowCollection rowCollection = new gov.energy.nbc.car.model.mongodb.common.RowCollection(dataUpload.columnNames, dataUpload.rowData);
-
-            IMetadata metadata = new Metadata(metadataJson);
-
-            PerformanceLogger performanceLogger = new PerformanceLogger(log, "new Dataset()");
-            DatasetDocument datasetDocument = new DatasetDocument(metadata);
-            performanceLogger.done();
-
-            ObjectId objectId = datasetDAO.add(datasetDocument, rowCollection);
-
-            return objectId.toHexString();
-        }
-        catch (IOException e) {
-            log.error(e);
-            throw new RuntimeException(e);
-        }
-    }
-
-    @Override
-    public m_DatasetDAO getDatasetDAO() {
-
-        return datasetDAO;
-    }
-}
