+++ src/main/java/gov/energy/nbc/car/bo/mongodb/singleCellSchemaApproach/r_RowBO.java
@@ -1,0 +1,19 @@
+package gov.energy.nbc.car.bo.mongodb.singleCellSchemaApproach;
+
+import gov.energy.nbc.car.bo.mongodb.AbsRowBO;
+import gov.energy.nbc.car.dao.mongodb.abandonedApproaches.everthingInTheRowCollectionApproach.r_RowDAO;
+import gov.energy.nbc.car.settings.ISettings;
+import org.apache.log4j.Logger;
+
+public class r_RowBO extends AbsRowBO {
+
+    protected Logger log = Logger.getLogger(getClass());
+
+    public r_RowBO(ISettings settings) {
+        super(settings);
+    }
+
+    public void init(ISettings settings) {
+        rowDAO = new r_RowDAO(settings);
+    }
+}
