--- src/main/java/gov/energy/nbc/car/dao/mongodb/dto/DataOrMetatdata.java
@@ -1,7 +1,0 @@
-//package gov.energy.nbc.car.dao.mongodb.dto;
-//
-//public enum DataOrMetatdata {
-//
-//    DATA,
-//    METADATA
-//}
