--- src/test/java/gov/energy/nbc/car/bo/mongodb/multipleCellSchemaApproach/m_RowBOTest.java
@@ -1,15 +1,0 @@
-package gov.energy.nbc.car.bo.mongodb.multipleCellSchemaApproach;
-
-import gov.energy.nbc.car.app.DataRepositoryApplication;
-import gov.energy.nbc.car.bo.mongodb.AbsRowBOTest;
-import gov.energy.nbc.car.settings.Settings;
-
-
-public class m_RowBOTest extends AbsRowBOTest
-{
-    @Override
-    protected DataRepositoryApplication createAppSingleton(Settings settings) {
-
-        return new DataRepositoryApplication(settings, new m_BusinessObjects(settings));
-    }
-}
