--- src/test/java/gov/energy/nbc/car/bo/mongodb/singleRowSchemaApproach/singleCellSchemaApproach/r_DatasetBOTest.java
@@ -1,15 +1,0 @@
-package gov.energy.nbc.car.bo.mongodb.singleRowSchemaApproach.singleCellSchemaApproach;
-
-import gov.energy.nbc.car.app.DataRepositoryApplication;
-import gov.energy.nbc.car.bo.mongodb.AbsDatasetBOTest;
-import gov.energy.nbc.car.settings.Settings;
-
-
-public class r_DatasetBOTest extends AbsDatasetBOTest
-{
-    @Override
-    protected DataRepositoryApplication createAppSingleton(Settings settings) {
-
-        return new DataRepositoryApplication(settings, new r_BusinessObjects(settings));
-    }
-}
