--- src/test/java/gov/energy/nbc/car/bo/mongodb/singleRowSchemaApproach/singleCellSchemaApproach/r_RowBOTest.java
@@ -1,15 +1,0 @@
-package gov.energy.nbc.car.bo.mongodb.singleRowSchemaApproach.singleCellSchemaApproach;
-
-import gov.energy.nbc.car.app.DataRepositoryApplication;
-import gov.energy.nbc.car.bo.mongodb.AbsRowBOTest;
-import gov.energy.nbc.car.settings.Settings;
-
-
-public class r_RowBOTest extends AbsRowBOTest
-{
-    @Override
-    protected DataRepositoryApplication createAppSingleton(Settings settings) {
-
-        return new DataRepositoryApplication(settings, new r_BusinessObjects(settings));
-    }
-}
