+++ src/test/java/gov/energy/nbc/car/dao/mongodb/abandonedApproaches/everthingInTheRowCollectionApproach/r_RowDAOTest.java
@@ -1,0 +1,15 @@
+package gov.energy.nbc.car.dao.mongodb.abandonedApproaches.everthingInTheRowCollectionApproach;
+
+import gov.energy.nbc.car.app.DataRepositoryApplication;
+import gov.energy.nbc.car.bo.mongodb.singleCellSchemaApproach.r_BusinessObjects;
+import gov.energy.nbc.car.dao.mongodb.AbsRowDAOTest;
+import gov.energy.nbc.car.settings.Settings;
+
+public class r_RowDAOTest extends AbsRowDAOTest {
+
+    @Override
+    protected DataRepositoryApplication createAppSingleton(Settings settings) {
+
+        return new DataRepositoryApplication(settings, new r_BusinessObjects(settings));
+    }
+}
