--- src/test/java/gov/energy/nbc/car/dao/mongodb/multipleCellCollectionsApproach/m_RowDAOTest.java
@@ -1,15 +1,0 @@
-package gov.energy.nbc.car.dao.mongodb.multipleCellCollectionsApproach;
-
-import gov.energy.nbc.car.app.DataRepositoryApplication;
-import gov.energy.nbc.car.bo.mongodb.multipleCellSchemaApproach.m_BusinessObjects;
-import gov.energy.nbc.car.dao.mongodb.AbsRowDAOTest;
-import gov.energy.nbc.car.settings.Settings;
-
-public class m_RowDAOTest extends AbsRowDAOTest {
-
-    @Override
-    protected DataRepositoryApplication createAppSingleton(Settings settings) {
-
-        return new DataRepositoryApplication(settings, new m_BusinessObjects(settings));
-    }
-}
